refactor: rename `vm` subsystem to `mem`
