+++ kernel/src/wasm/mod.rs
@@ -24,9 +24,9 @@ mod type_registry;
 mod utils;
 mod values;
 
+use crate::mem::ArchAddressSpace;
+use crate::mem::frame_alloc::FRAME_ALLOC;
 use crate::scheduler::scheduler;
-use crate::vm::ArchAddressSpace;
-use crate::vm::frame_alloc::FRAME_ALLOC;
 use crate::{enum_accessors, owned_enum_accessors};
 use core::fmt::Write;
 use wasmparser::Validator;
